sua lai databse và model

+++ bookingtrain/src/main/java/com/example/bookingtrain/repository/PermissionRepository.java
@@ -3,6 +3,6 @@ package com.example.bookingtrain.repository;
 import com.example.bookingtrain.model.Permission;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PermissionRepository extends JpaRepository<Permission, Long> {
+public interface PermissionRepository extends JpaRepository<Permission, Integer> {
 
 }
